feat: change entries printing system and add added by
in the printing system, you'll should now be able to see whon added
this, in orded to better indentify who did this change, and also make
the type be the status. this was a request by Thais, in order to make
the system more concise and able to quickly identify the type of
docuemnt.